Check whether the IPC Acute Malnutrition sample size requirements were met

Verify whether the minimum sample size requirements for the area of analysis were met, in accordance with the IPC Acute Malnutrition (IPC AMN) protocols.

Usage

check_sample_size(df, .group, .data_type = c("survey", "screening", "ssite"))

Arguments

df

A dataset of class data.frame to check.

.group

A vector of class integer of the cluster ID's for survey, screening or site ID's for screenings and sentinel sites.

.data_type

A choice between "survey" for survey data, "screening" for screening data or "ssite" for community-based sentinel site data.

Value

A summarised table of three columns: groups for the total number of unique cluster or screening or site IDs; n_obs for the respective total number of children; and meet_ipc for whether the IPC AMN requirements were met.

Details

The IPC Manual.

Examples

anthro.01 |>
dplyr::group_by(area) |>
check_sample_size(
.group = cluster,
.data_type = "survey"
)
#> # A tibble: 2 × 4
#>   area       groups n_obs meet_ipc
#>   <chr>       <int> <int> <chr>   
#> 1 District E     28   505 yes     
#> 2 District G     30   686 yes