Get the overall acceptability score from the acceptability classification scores

Calculate the total amount of penalty points based on each plausibility test result acceptability classification for WFHZ and MFAZ.

Usage

compute_quality_score(df, type = c("mfaz", "whz"))

Arguments

df

A dataset object of class data.frame to calculate from.

type

A choice between "wfhz" and "mfaz" for the basis on which the calculations should be made.

Value

A data.frame based on df with a new column named "quality_score" for the overall of acceptability (of quality) score.

Examples

## A sample data ----

df <- data.frame(
flagged_class = "Excellent",
age_ratio_class = "Good",
sex_ratio_class = "Problematic",
dps_class = "Excellent",
sd_class = "Excellent",
skew_class = "Good",
kurt_class = "Acceptable"
)

## Apply the function ----
compute_quality_score(df, type = "mfaz")
#> [1] 16