Check the plausibility and acceptability of MUAC-for-age z-score (MFAZ) data

Check the overall plausibility and acceptability of MFAZ data through a structured test suite encompassing checks for sampling and measurement-related biases in the dataset. This test suite follows the recommendation made by Bilukha & Kianian (2023) on the plausibility of constructing a comprehensive plausibility check for MUAC data similar to weight-for-height z-score to evaluate its acceptability when age values are available in the dataset.

The function works on a data.frame returned from wrangling functions for age and for MUAC-for-age z-score data available from this package.

Usage

mw_plausibility_check_mfaz(df, sex, muac, age, flags)

Arguments

df

A data.frame object to check.

sex

A numeric vector for child's sex with 1 = males and 2 = females.

muac

A numeric vector of child's MUAC in centimeters.

age

A vector of class double of child's age in months.

flags

A numeric vector of flagged records.

Value

A single row summary tibble with 17 columns containing the plausibility check results and their respective acceptability ratings.

Details

Whilst the function uses the same checks and criteria as those for weight-for-height z-scores in the SMART plausibility check, the percent of flagged records is evaluated using different cut-off points, with a maximum acceptability of 2.0% as shown below:

Excellent	Good	Acceptable	Problematic

0.0 - 1.0	>1.0 - 1.5	>1.5 - 2.0	>2.0

References

Bilukha, O., & Kianian, B. (2023). Considerations for assessment of measurement quality of mid‐upper arm circumference data in anthropometric surveys and mass nutritional screenings conducted in humanitarian and refugee settings. Maternal & Child Nutrition, 19, e13478. https://doi.org/10.1111/mcn.13478

SMART Initiative (2017). Standardized Monitoring and Assessment for Relief and Transition. Manual 2.0. Available at: https://smartmethodology.org.

See also

Examples

## First wrangle age data ----
data <- mw_wrangle_age(
  df = anthro.01,
  dos = dos,
  dob = dob,
  age = age,
  .decimals = 2
)

## Then wrangle MUAC data ----
data_muac <- mw_wrangle_muac(
  df = data,
  sex = sex,
  age = age,
  muac = muac,
  .recode_sex = TRUE,
  .recode_muac = TRUE,
  .to = "cm"
)
#> ================================================================================

## And finally run plausibility check ----
mw_plausibility_check_mfaz(
  df = data_muac,
  flags = flag_mfaz,
  sex = sex,
  muac = muac,
  age = age
)
#> # A tibble: 1 × 17
#>       n flagged flagged_class sex_ratio sex_ratio_class age_ratio
#>   <int>   <dbl> <fct>             <dbl> <chr>               <dbl>
#> 1  1191 0.00504 Excellent         0.297 Excellent           0.636
#> # ℹ 11 more variables: age_ratio_class <chr>, dps <dbl>, dps_class <chr>,
#> #   sd <dbl>, sd_class <chr>, skew <dbl>, skew_class <fct>, kurt <dbl>,
#> #   kurt_class <fct>, quality_score <dbl>, quality_class <fct>