Check the plausibility of the data

Verify the overall acceptability of the data through a set of structured tests around sampling and measurement-related biases in the data.

Usage

check_plausibility_mfaz(df, sex, muac, age, flags, area)

check_plausibility_wfhz(df, sex, age, weight, height, flags, area)

check_plausibility_muac(df, flags, sex, muac)

Arguments

df

A dataset object of class data.frame to check. It should have been wrangled using this package's wranglers.

sex

A vector of class numeric of child's sex: 1 for boy and 2 for girl.

muac

A vector of class double of child's MUAC in centimeters.

age

A vector of class double of child's age in months.

flags

A vector of class numeric of flagged observations.

area

A vector of class character of the geographical location where data was collected and for which the analysis should be aggregated.

weight

A vector of class double of child's weight in kilograms.

height

A vector of class double of child's height in centimeters.

Value

A summarised data.frame of plausibility test results and their respective acceptability ratings.

Examples

## Check the plausibility of WFHZ data ----

anthro.01 |>
process_age(
svdate = "dos",
birdate = "dob",
age = age
) |>
process_wfhz_data(
sex = sex,
weight = weight,
height = height,
.recode_sex = TRUE
) |>
check_plausibility_wfhz(
sex = sex,
age = age,
weight = weight,
height = height,
flags = flag_wfhz,
area = area
)
#> ================================================================================
#> # A tibble: 2 × 20
#>   area           n flagged flagged_class sex_ratio sex_ratio_class age_ratio
#>   <chr>      <int>   <dbl> <fct>             <dbl> <chr>               <dbl>
#> 1 District E   505 0.00792 Excellent         0.593 Excellent           0.265
#> 2 District G   686 0.0117  Excellent         0.380 Excellent           0.925
#> # ℹ 13 more variables: age_ratio_class <chr>, dps_wgt <dbl>,
#> #   dps_wgt_class <chr>, dps_hgt <dbl>, dps_hgt_class <chr>, sd <dbl>,
#> #   sd_class <chr>, skew <dbl>, skew_class <fct>, kurt <dbl>, kurt_class <fct>,
#> #   quality_score <dbl>, quality_class <fct>

## Check the plausibility of MFAZ data ----

anthro.01 |>
process_age(
svdate = "dos",
birdate = "dob",
age = age
) |>
process_muac_data(
sex = sex,
age = "age",
muac = muac,
.recode_sex = TRUE,
.recode_muac = TRUE,
unit = "cm"
) |>
check_plausibility_mfaz(
flags = flag_mfaz,
sex = sex,
muac = muac,
age = age,
area = area
)
#> ================================================================================
#> # A tibble: 2 × 18
#>   area           n flagged flagged_class sex_ratio sex_ratio_class age_ratio
#>   <chr>      <int>   <dbl> <fct>             <dbl> <chr>               <dbl>
#> 1 District E   505 0       Excellent         0.593 Excellent           0.126
#> 2 District G   686 0.00875 Excellent         0.380 Excellent           0.490
#> # ℹ 11 more variables: age_ratio_class <chr>, dps <dbl>, dps_class <chr>,
#> #   sd <dbl>, sd_class <chr>, skew <dbl>, skew_class <fct>, kurt <dbl>,
#> #   kurt_class <fct>, quality_score <dbl>, quality_class <fct>

## Check the plausibility of the absolute MUAC values ----

anthro.01 |>
process_muac_data(
sex = sex,
muac = muac,
age = NULL,
.recode_sex = TRUE,
.recode_muac = FALSE,
unit = "none"
) |>
check_plausibility_muac(
flags = flag_muac,
sex = sex,
muac = muac
)
#> # A tibble: 1 × 9
#>       n flagged flagged_class sex_ratio sex_ratio_class   dps dps_class    sd
#>   <int>   <dbl> <fct>             <dbl> <chr>           <dbl> <chr>     <dbl>
#> 1  1191 0.00252 Excellent         0.297 Excellent        5.39 Excellent  11.1
#> # ℹ 1 more variable: sd_class <fct>