Score the acceptability classification of the standard deviation and percentage of flagged data test results

Attribute a penalty point based on the acceptability classification in which the plausibility test result falls.

Usage

assign_penalty_points_flags_and_sd(x)

assign_penalty_points_age_sex_ratio(x)

assign_penalty_points_skew_kurt(x)

Arguments

x

A vector of class character of acceptability classification of the plausibility test results.

Value

A vector of class integer of the same length as x for the score.

Details

The scoring criteria is as in SMART Plausibility checks.