A helper function to determine the MUAC prevalence analysis approach to follow

It determines the analysis approach to follow for a given analysis area on the basis of the rate of acceptability of the age ratio test and the standard deviation analysis result.

Usage

tell_muac_analysis_strategy(age_ratio_class, sd_class)

Arguments

age_ratio_class

A vector of class character of the acceptability classification of the age ratio test result.

sd_class

A vector of class character of the acceptability classification of the standard deviation analysis result.

Value

A vector of class character of the same length as the input vectors, containing values indicating the analysis approach for each analysis area: "weighted", "unweighted" and "missing".

Details

When "weighted", the CDC weighting approach is applied to correct for age bias; when "unweighted" a normal complex sample analysis is applied; when "missing" NA gets thrown.