Wrangle weight-for-height and MUAC data

This function performs data wrangling by calculating weight-for-height and MUAC-for-age z-scores, followed by the detection and flagging of outliers. For MUAC data, if age is not supplies, z-scores do not get computed. In such cases, outlier detection and flagging are based on the absolute MUAC values.

Usage

process_wfhz_data(df, sex, weight, height, .recode_sex = TRUE)

process_muac_data(
  df,
  sex,
  muac,
  age = NULL,
  .recode_sex = TRUE,
  .recode_muac = TRUE,
  unit = c("cm", "mm", "none")
)

Arguments

df

A dataset of class data.frame to wrangle data from.

sex

A numeric or character vector of child's sex. Code values should be 1 or "m" for boy and 2 or "f" for girl. The variable name must be sex, otherwise it will not work.

weight

A vector of class double of child's weight in kilograms.

height

A vector of class double of child's height in centimeters.

.recode_sex

Logical. Default is FALSE. Setting to TRUE assumes that the sex variable is a character vector of values "m" for boys and "f" for girls and will recode them to 1 and 2 respectively.

muac

A vector of class double or integer of the absolute MUAC values.

age

A double vector of child's age in months. It must be named age, otherwise it will not work.

.recode_muac

Logical. Default is FALSE. Set to TRUE if MUAC values should be converted to either centimeters or millimeters.

unit

A choice of the unit to which the MUAC values should be converted. "cm" for centimeters, "mm" for millimeters and "none" to leave as it is.

Value

A data frame based on df. New variables named wfhz and flag_wfhz, of child's weight-for-height z-scores and flags, or mfaz and flag_mfaz, of child's MUAC-for-age z-scores and flags, will be created. For MUAC, when age is not supplied only flag_muac variable is created. This refers to flags based on the absolute MUAC values as recommended by Bilukha, O., & Kianian, B. (2023)..

Details

The flagging criterion used for the WFHZ and MFAZ is as in SMART plausibility check. A fixed flagging criterion is used for the absolute MUAC values. This is as recommended by Bilukha, O., & Kianian, B. (2023).

Examples

## An example application of `process_wfhz_data()` ----

anthro.01 |>
process_wfhz_data(
sex = sex,
weight = weight,
height = height,
.recode_sex = TRUE
)
#> ================================================================================
#> # A tibble: 1,191 × 13
#>    area      dos        cluster  team   sex dob      age weight height edema
#>    <chr>     <date>       <int> <int> <dbl> <date> <int>  <dbl>  <dbl> <chr>
#>  1 District… 2023-12-04       1     3     1 NA        59   15.6  109.  n    
#>  2 District… 2023-12-04       1     3     1 NA         8    7.5   68.6 n    
#>  3 District… 2023-12-04       1     3     1 NA        19    9.7   79.5 n    
#>  4 District… 2023-12-04       1     3     2 NA        49   14.3  100.  n    
#>  5 District… 2023-12-04       1     3     2 NA        32   12.4   92.1 n    
#>  6 District… 2023-12-04       1     3     2 NA        17    9.3   77.8 n    
#>  7 District… 2023-12-04       1     3     2 NA        20   10.1   80.4 n    
#>  8 District… 2023-12-04       1     3     2 NA        27   11.7   87.1 n    
#>  9 District… 2023-12-04       1     3     1 NA        46   13.6   98   n    
#> 10 District… 2023-12-04       1     3     1 NA        58   17.2  109.  n    
#> # ℹ 1,181 more rows
#> # ℹ 3 more variables: muac <int>, wfhz <dbl>, flag_wfhz <dbl>

## An example application of `process_muac_data()` ----

### Sample data ----
df <- data.frame(
 survey_date = as.Date(c(
 "2023-01-01", "2023-01-01", "2023-01-01", "2023-01-01", "2023-01-01")),
 birthdate = as.Date(c(
 "2019-01-01", NA, "2018-03-20", "2019-11-05", "2021-04-25")),
 age = c(NA, 36, NA, NA, NA),
 sex = c("m", "f", "m", "m", "f"),
 muac = c(110, 130, 300, 123, 125)
 )

 ### The application of the function ----

 df |>
 process_age(
 svdate = "survey_date",
 birdate = "birthdate",
 age = age
 ) |>
 process_muac_data(
 sex = sex,
 age = "age",
 muac = muac,
 .recode_sex = TRUE,
 .recode_muac = TRUE,
 unit = "cm"
 )
#> ================================================================================
#> # A tibble: 5 × 8
#>   survey_date birthdate    age   sex  muac age_days  mfaz flag_mfaz
#>   <date>      <date>     <dbl> <dbl> <dbl>    <dbl> <dbl>     <dbl>
#> 1 2023-01-01  2019-01-01  48       1  11      1461. -4.66         1
#> 2 2023-01-01  NA          36       2  13      1096. -2.25         0
#> 3 2023-01-01  2018-03-20  57.4     1  30      1748.  7.43         1
#> 4 2023-01-01  2019-11-05  37.9     1  12.3    1153. -3.21         0
#> 5 2023-01-01  2021-04-25  20.2     2  12.5     616. -1.95         0