Estimate the prevalence of wasting based on MUAC for non-survey data

It is common to estimate prevalence of wasting from non survey data, such as screenings or any other community-based surveillance systems. In such situations, the analysis usually consists only in estimating the point prevalence and the counts of positive cases, without necessarily estimating the uncertainty. This function serves this use.

The quality of the data is first evaluated by calculating and rating the standard deviation of MFAZ and the p-value of the age ratio test. Prevalence is calculated only when the standard deviation of MFAZ is not problematic. If both standard deviation of MFAZ and p-value of age ratio test is not problematic, straightforward prevalence estimation is performed. If standard deviation of MFAZ is not problematic but p-value of age ratio test is problematic, age-weighting is applied to prevalence estimation to account for the over-representation of younger children in the sample. If standard deviation of MFAZ is problematic, no estimation is done and an NA value is returned. Outliers are detected based on SMART flagging criteria for MFAZ. Identified outliers are then excluded before prevalence estimation is performed.

Usage

mw_estimate_prevalence_screening(df, muac, edema = NULL, .by = NULL)

Arguments

df

A tibble object produced by mw_wrangle_muac() and mw_wrangle_age() functions. Note that MUAC values in df must be in millimeters unit after using mw_wrangle_muac(). Also, df must have a variable called cluster which contains the primary sampling unit identifiers.

muac

A numeric or integer vector of raw MUAC values. The measurement unit of the values should be millimeters.

edema

A character vector for presence of nutritional edema coded as "y" for presence of nutritional edema and "n" for absence of nutritional edema. Default is NULL.

.by

A character or numeric vector of the geographical areas or identifiers for where the data was collected and for which the analysis should be summarised for.

Value

A summary tibble for the descriptive statistics about combined wasting.

References

SMART Initiative (no date). Updated MUAC data collection tool. Available at: https://smartmethodology.org/survey-planning-tools/updated-muac-tool/

See also

Examples

mw_estimate_prevalence_screening(
  df = anthro.02,
  muac = muac,
  edema = edema,
  .by = province
)
#> # A tibble: 2 × 7
#>   province gam_n  gam_p sam_n   sam_p mam_n  mam_p
#>   <chr>    <dbl>  <dbl> <dbl>   <dbl> <dbl>  <dbl>
#> 1 Nampula     61 0.0590    19 0.0184     42 0.0406
#> 2 Zambezia    57 0.0500    10 0.00876    47 0.0412

## With `edema` set to `NULL` ----
mw_estimate_prevalence_screening(
  df = anthro.02,
  muac = muac,
  edema = NULL,
  .by = province
)
#> # A tibble: 2 × 7
#>   province gam_n  gam_p sam_n   sam_p mam_n  mam_p
#>   <chr>    <dbl>  <dbl> <dbl>   <dbl> <dbl>  <dbl>
#> 1 Nampula     53 0.0513    10 0.00967    43 0.0416
#> 2 Zambezia    53 0.0465     6 0.00526    47 0.0412

## With `.by` set to `NULL` ----
mw_estimate_prevalence_screening(
  df = anthro.02,
  muac = muac,
  edema = NULL,
  .by = NULL
)
#> # A tibble: 1 × 6
#>   gam_n  gam_p sam_n   sam_p mam_n  mam_p
#>   <dbl>  <dbl> <dbl>   <dbl> <dbl>  <dbl>
#> 1   106 0.0487    16 0.00736    90 0.0414