Check the plausibility and acceptability of raw MUAC data

Check the overall plausibility and acceptability of raw MUAC data through a structured test suite encompassing sampling and measurement-related biases checks in the data set. The test suite in this function follows the recommendation made by Bilukha, O., & Kianian, B. (2023).

Usage

mw_plausibility_check_muac(df, sex, muac, flags)

Arguments

df

An object of class data.frame to check. It must have been wrangled using this package's wrangling function for MUAC.

sex

A vector of class numeric of child's sex.

muac

A vector of class double of child's MUAC in centimeters.

flags

A vector of class numeric of flagged records.

Value

A summarized table of class data.frame, of length 9 and width 1, for the plausibility test results and their respective acceptability ratings.

Details

Cut-off points used for the percent of flagged records:

Excellent	Good	Acceptable	Problematic

0.0 - 1.0	>1.0 - 1.5	>1.5 - 2.0	>2.0

References

Bilukha, O., & Kianian, B. (2023). Considerations for assessment of measurement quality of mid‐upper arm circumference data in anthropometric surveys and mass nutritional screenings conducted in humanitarian and refugee settings. Maternal & Child Nutrition, 19, e13478. https://doi.org/10.1111/mcn.13478

SMART Initiative (2017). Standardized Monitoring and Assessment for Relief and Transition. Manual 2.0. Available at: https://smartmethodology.org.

See also

Examples

## First wranlge MUAC data ----
df_muac <- mw_wrangle_muac(
  df = anthro.01,
  sex = sex,
  muac = muac,
  age = NULL,
  .recode_sex = TRUE,
  .recode_muac = FALSE,
  .to = "none"
)

## Then run the plausibility check ----
mw_plausibility_check_muac(
  df = df_muac,
  flags = flag_muac,
  sex = sex,
  muac = muac
)
#> # A tibble: 1 × 9
#>       n flagged flagged_class sex_ratio sex_ratio_class   dps dps_class    sd
#>   <int>   <dbl> <fct>             <dbl> <chr>           <dbl> <chr>     <dbl>
#> 1  1191 0.00252 Excellent         0.297 Excellent        5.39 Excellent  11.1
#> # ℹ 1 more variable: sd_class <fct>