Identify, flag outliers and remove them

Identify outlier z-scores for weight-for-height (WFHZ) and MUAC-for-age (MFAZ) following the SMART methodology. The function can also be used to detect outliers for height-for-age (HFAZ) and weight-for-age (WFAZ) z-scores following the same approach.

For raw MUAC values, outliers constitute values that are less than 100 millimeters or greater than 200 millimeters.

Removing outliers consist in setting the outlier record to NA and not necessarily to delete it from the dataset. This is useful in the analysis procedures where outliers must be removed, such as the analysis of the standard deviation.

Usage

flag_outliers(x, .from = c("zscores", "raw_muac"))

remove_flags(x, .from = c("zscores", "raw_muac"))

Arguments

x

A vector of class numeric of WFHZ, MFAZ, HFAZ, WFAZ or raw MUAC values. The latter should be in millimeters. If the class is different than expected, the function will stop execution and return an error message indicating the type of mismatch.

.from

A choice between zscores and raw_muac for where outliers should be detected and flagged from.

Value

A vector of the same length as x for flagged records coded as 1 for is a flag and 0 not a flag.

Details

For z-score-based detection, flagged records represent outliers that deviate substantially from the sample's z-score mean, making them unlikely to reflect accurate measurements. For raw MUAC values, flagged records are those that fall outside the acceptable fixed range. Including such outliers in the analysis could compromise the accuracy and precision of the resulting estimates.

The flagging criterion used for raw MUAC values is based on a recommendation by Bilukha, O., & Kianian, B. (2023).

References

Bilukha, O., & Kianian, B. (2023). Considerations for assessment of measurement quality of mid‐upper arm circumference data in anthropometric surveys and mass nutritional screenings conducted in humanitarian and refugee settings. Maternal & Child Nutrition, 19, e13478. Available at https://doi.org/10.1111/mcn.13478

SMART Initiative (2017). Standardized Monitoring and Assessment for Relief and Transition. Manual 2.0. Available at: https://smartmethodology.org.

Examples

## Sample data of raw MUAC values ----
x <- anthro.01$muac

## Apply the function with `.from` set to "raw_muac" ----
flag_outliers(x, .from = "raw_muac")
#>    [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>   [38] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>   [75] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [112] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [149] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [186] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [223] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [260] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [297] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [334] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [371] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [408] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [445] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [482] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [519] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [556] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [593] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [630] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [667] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [704] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [741] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [778] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [815] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [852] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [889] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [926] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [963] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [1000] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [1037] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [1074] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [1111] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0
#> [1148] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [1185] 0 0 0 0 0 0 0

## Sample data of z-scores (be it WFHZ, MFAZ, HFAZ or WFAZ) ----
x <- anthro.02$mfaz

# Apply the function with `.from` set to "zscores" ----
flag_outliers(x, .from = "zscores")
#>    [1]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 NA  0  0 NA
#>   [25]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>   [49]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>   [73]  0  0  0  0  1  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>   [97]  0  0  0  0  0  0  0  0  0  0  0  0  0 NA  0  0  0  0  0  0  0  0  0  1
#>  [121]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [145]  0  0 NA  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  1
#>  [169]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [193]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [217]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [241]  0  0  0  0  0  0  0  0  0  0  0  1  0  0  0  0  0  0  0  0  0 NA  0  1
#>  [265]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [289]  0  0  0  0  0  0  0  0  0  0  0  1 NA  0  1  0  0  0  1  0  0  0  0  0
#>  [313]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [337]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [361]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [385]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [409]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [433]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [457]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [481]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 NA
#>  [505]  0  0  0 NA  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [529]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 NA  0  0  0
#>  [553]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [577]  0  0  0  0  0  0 NA  0  0  0  0  0  0  1  0  0  0  0  0  0  0  0  0  0
#>  [601] NA  0  0  0  0  0 NA  0  0  0  0  0  0  1  0  0  0  0  0  0  0  0  0  0
#>  [625]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [649]  0  0  0  0  0  0  0  0  0 NA  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [673]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [697]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [721]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [745] NA  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [769]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [793]  0  0  0  0  0  0  0  0  0  0 NA  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [817]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [841]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [865]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [889] NA  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [913]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [937]  0  0  0  0  0  0  0  0  0  0  0  0  0  0 NA  0  0  0  0  0  0  0  0  0
#>  [961]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>  [985]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [1009]  0  0  0 NA  0  0  0  0 NA NA  0  0  0  0 NA  0  0  0  0  0  0 NA  0  0
#> [1033]  0  0  0  0  0  0 NA  0  0  0  0  0  0  0 NA NA  0  0  0  0  0  0  0  0
#> [1057]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 NA  0  0  0  0  0  0  0  0
#> [1081]  0  0  0  0 NA  0  0  0  0  0  0  0  0  0  0  0  0 NA  0  0  0  0  0  0
#> [1105]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 NA  0  0  0  0  0
#> [1129]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [1153]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 NA  0  0  0  0  0
#> [1177]  0  0  0  0  0  0  0  0  0  0 NA  0  0  0  0  0 NA  0  0  0  0  0  0  0
#> [1201]  0  0  0 NA  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [1225]  0  0  0  0  0  0  0  0  0  0  0  0  0  0 NA  0  0  0  0  0  0  0  0  0
#> [1249]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [1273]  0  0 NA  0  0  0  0  0  0  0  0  1  0  0  0  0  0  0  0  0  0  0  0  0
#> [1297]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [1321]  0  0  0  0  0  0  0  1  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [1345]  0  0  0  0 NA  0  0  0  0  0  0  0  0  0  0  0  0  0 NA  0  0  0  0  0
#> [1369]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [1393]  0  0  0  0  0  0  0  0  0  0 NA  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [1417]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [1441]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [1465]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [1489]  0  0  0  0  0 NA  0  0  0  0 NA  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [1513]  0  1  0  0  0  0  0  0  0  0  0  0  0  0  1  0  0  0  0  0  0  0  0  0
#> [1537]  0 NA  0  0  0  1  0  0  0  0  0  0  0  0  0  0  0  0  0  0 NA  0  0  0
#> [1561]  0  0  0  0  1  0  0  0  0  0  0  0  0  0  0  0  1  0  0  0  0  0  0  0
#> [1585]  0  0  0  0  0  0  0  0  0  0 NA  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [1609]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 NA  0 NA  0  0  0  0
#> [1633]  0  0  0  0  0  0 NA  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [1657]  0  0  0  0 NA  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [1681]  1  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  1  0
#> [1705]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [1729]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [1753]  0  0  0  0  0 NA  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 NA  0  0
#> [1777]  0  0 NA  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [1801] NA NA  0  0 NA  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [1825]  0  0  0  0  1  0  0  0 NA  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [1849]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [1873]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 NA  0  0  0 NA  0  0  0  0
#> [1897]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 NA  0  0
#> [1921]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [1945]  0  0 NA  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [1969]  0 NA  0  0 NA  0  0  0  0  0  0  0  0  0  0  0  0  0 NA  0  0  0  0  0
#> [1993]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [2017]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [2041]  0  0  0  0  0  0  0  0  0  0  0 NA  0  0  0  0  0  0 NA  0  0  0  0  0
#> [2065]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [2089]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [2113]  0  0  0  0  0  0  0  0  0  0 NA  0  0 NA  0  0  0 NA  0  0  0 NA  0  0
#> [2137]  0 NA  0  0  0  0  0  0  0 NA  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [2161]  0  0  0  0  0  0  0  0  0  0 NA  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [2185]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [2209]  0  0  0  0  0  0  0  0  0  0  0  0  0 NA  0  0  0  0  0  0  0  0  0  0
#> [2233]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> [2257]  0  0  0  0 NA  0  0  0  0  0  0

## With `.from` set to "zscores" ----
remove_flags(
  x = wfhz.01$wfhz,
  .from = "zscores"
)
#>   [1]  1.833  0.278 -0.123  1.442  0.652  0.469  0.886 -0.701  0.232 -0.384
#>  [11] -0.124  0.359 -0.888 -0.879 -1.818  0.502  0.862 -0.321  0.016 -0.742
#>  [21] -0.589  1.220 -2.791  0.859 -2.932 -0.479  0.177  1.547  0.491 -0.278
#>  [31] -0.534  0.177 -0.756 -0.221  0.584  0.723  0.560 -1.286 -0.814  0.227
#>  [41] -0.361 -1.154  0.010 -1.170 -0.377 -1.254 -0.244  0.524  0.495  0.488
#>  [51] -2.749 -1.290  0.350 -2.784 -1.402 -0.943  0.333 -0.176 -2.522 -1.870
#>  [61] -2.142     NA -2.173 -0.947     NA -1.256  0.439 -0.169  0.168 -0.839
#>  [71] -0.705 -1.054  1.284  0.064 -0.465  0.445 -0.700  0.728 -1.608 -0.742
#>  [81] -1.406 -0.502 -0.544  0.581 -0.414 -0.397 -0.532 -0.648 -1.776 -2.669
#>  [91] -2.129 -2.404  0.217 -0.947 -0.520 -0.320  0.660  0.757  1.562  0.159
#> [101] -0.686  0.705  0.276 -1.124  0.378 -0.719  0.349 -0.644  0.045  0.350
#> [111]  0.582 -2.617  1.065  0.718  0.200 -0.604  0.597  0.529  0.097 -3.112
#> [121]     NA  0.204 -0.456  0.581 -1.340 -0.950  0.243  0.863     NA -1.050
#> [131]  2.430 -0.314  0.446 -0.303 -0.515  0.402  1.171 -1.577 -1.912 -0.380
#> [141]  0.694 -0.779 -1.794  0.660  1.815  0.376  0.016  0.628 -2.090 -0.198
#> [151] -1.339  0.681 -1.655     NA  0.721 -1.025 -0.912  0.916  1.946  0.074
#> [161]  0.178  0.052  0.131  0.566 -0.409  0.286 -1.999 -0.527  0.977  0.076
#> [171] -0.524  1.346 -0.410 -0.716 -0.271  1.354 -2.332 -0.403 -0.735  1.429
#> [181]  1.090 -0.151 -0.795 -2.563 -0.308 -1.369 -2.206  0.314 -1.107 -0.961
#> [191]  0.500 -1.474 -0.707  0.054 -2.704 -2.215 -0.726 -2.696 -1.320 -0.193
#> [201] -1.129  0.827 -1.533 -1.146 -0.915 -0.858 -1.001 -1.724  1.321  1.095
#> [211]  0.146  0.568  0.039  0.528 -0.236  0.812 -0.877 -0.201 -0.604 -2.323
#> [221]  0.056  0.827  0.557 -1.400  1.174 -0.687 -0.244  0.402  1.929 -0.968
#> [231] -1.391  0.520 -0.901 -1.168 -0.875 -0.350 -0.414 -0.412 -0.296  0.174
#> [241] -0.896 -0.155 -0.808  0.002  0.606 -2.295 -0.058 -0.656 -0.500  0.442
#> [251]  0.122 -0.712 -0.903  0.231 -0.635 -0.238  0.133 -1.232 -0.148 -0.794
#> [261]  0.135  0.401 -0.807  0.472  0.272 -0.662 -0.402 -1.039 -0.076 -0.876
#> [271] -0.703  0.527 -1.335  0.411 -1.861 -1.204 -2.173 -1.044 -0.495 -1.268
#> [281] -1.084 -1.251  0.649 -0.850  1.666  0.545  2.588  0.119 -1.211 -0.489
#> [291]  0.525  0.458  0.083  0.588  1.239 -0.954  0.008  1.269 -0.878  0.222
#> [301] -1.650 -1.103 -0.613

## With `.from` set to "raw_muac" ----
remove_flags(
  x = mfaz.01$muac,
  .from = "raw_muac"
)
#>   [1] 134 153 132 144 150 174 156 125 146 150 130 131 146 142 162 145 145 133
#>  [19] 135 111 142 125 165 154 162 150 154 147 161 127 138 175 180 134 154 140
#>  [37] 128 150 153 133 137 140 120 150 150 156 155 152 137 115 130 150 156 140
#>  [55] 133 130 153 120 130 124 153 143 144 160 130 152 130 118 115 146 115 130
#>  [73] 150 150 152 133 140 134 135 144 125 145 137 165 140 140 152 148 158 140
#>  [91] 145 145 140 136 137 136 145 128 145 130 116 152 160 141 112 149 132 130
#> [109] 135 134 130 142 150 138 140 134 131 145 132 113 138 123 129 150 150 136
#> [127] 145 145 130 130 136 183 148 160 132 140 162 128 160 160 143 142 139 135
#> [145] 153 149 147 147 145 134 170 140 170 144 127 125 143 153 142 142 150 160
#> [163] 135 140 145 128 151 114 105 130 100 141 154 125 120 128 152 148 134 158
#> [181] 152 137 127 152 138 155 150 143 121 152 134 135 148 123 120 143 149 151
#> [199] 118 135 121 151 138 135 139 118 145 120 143 148 126 140 128 141 157 116
#> [217] 164 149 145 146 131 149 159 158 131 142 158 140 134 130 144 123 160 140
#> [235] 140 165 130 137 137 124 144 134 153 144 138 130 115 154 136 138 136 138
#> [253] 139 141 132 143 127 145 136 141 144 185 129 121 149 161 117 145 120 180
#> [271] 121 133 131 149 140 139 135 187 159 154 149 141 138 147 133 131 153 145
#> [289] 136 164 149 135 149 137 148 136 147 139 115 149 132 129 160 156 150 150
#> [307] 150 135 150 141 140 150 140 160 150 128 140 122 141 139 142 120 141 113
#> [325] 133 143 122 131 121 121 130 121 117 123 112 138 123 133 133 132 112 132
#> [343] 112 131 122 141 121 121 106 131 141 153 132 149 134 139 131 140 154 152
#> [361] 139 141 131 131 142 151 143 156 142 153 156 142 132 153 150 190 175 135
#> [379] 170 140 140 170 164 151 170 180 148 127 146 145 148 130 163 150 140 145
#> [397] 140 175 190 120 146 159 131 129 158 131 167 156 144 117 144 165 140 165
#> [415] 140 138 158 140 153 137 147 135 175 130 151 138 141 140 182 153 130 130
#> [433] 150 193 120 121 140 137 141 136 146 150 131 123 168 123 132 160 137 182
#> [451] 134 153 130 180 150 143 133 140 129 151 130 148 141 161 149 148 155 139
#> [469] 114 129 162 139 145 132 131 149 136 142 147 137 154 136 138 153 134 142
#> [487] 144 131 145 144 147 128 146 143 145 142 140 161 133 147 117 120 153 161
#> [505] 158 152 143 165 175 138 141 154 149 141 120 148 160 145 160 138 160 135
#> [523] 143 148 153 120 130 149 140 143 156 148 135 144 144 158 160 148 133 142
#> [541] 126 144 129 153 130 161 120 140 135 145 140 139 160 141 160 121 150 168
#> [559] 147 131 135 138 128 140 144 165 145 141 150 136 150 141 149 141 124 125
#> [577] 153 142 140 195 120 123 185 120 143 175 197 179 183 151 142 145 150 143
#> [595] 137 129 139 130 140 138 142 140 151 130 153 156 148 140 158 165 142 140
#> [613] 145 150 124 160 160 158 130 148 120 155 139 141 158 130 135 156 142 143
#> [631] 159 130 160 167 153 155 135 149 142 160 143 157 148 156 153 145 151 138
#> [649] 137 142 139 160 136 152 123 149 130 168 175 161 139 146 143 138 153 158
#> [667] 147