Score the acceptability rating of the check results that constitutes the plausibility check suite

Attribute a score, also known as penalty point, for a given rate of acceptability of the standard deviation, proportion of flagged records, age and sex ratio, skewness, kurtosis and digit preference score check results.

The scoring criteria and thresholds follows the standards in the SMART plausibility check.

Usage

score_std_flags(x)

score_agesexr_dps(x)

score_skewkurt(x)

Arguments

x

A vector of class character containing the acceptability rate of a given test check. If the class does not match the expected type, the function will stop execution and return an error message indicating the type of mismatch.

Value

A vector of class integer of the same length as x for the acceptability score.

References

SMART Initiative (2017). Standardized Monitoring and Assessment for Relief and Transition. Manual 2.0. Available at: https://smartmethodology.org.